refactor: 테스트 태그 제거 및 `ExpireDateStatus` 위치 변경
