+++ app/src/main/java/woowacourse/payments/ui/features/addcard/components/CardNumberField.kt
@@ -8,7 +8,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
@@ -16,7 +15,6 @@ import woowacourse.payments.R
 import woowacourse.payments.ui.components.AppTextField
 import woowacourse.payments.ui.features.addcard.CardUiState
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
-import woowacourse.payments.ui.util.AppTestTags.CARD_NUMBER_FIELD
 import woowacourse.payments.ui.util.CardNumberVisualTransformation
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -29,7 +27,7 @@ fun CardNumberField(
     AppTextField(
         value = value,
         onValueChange = onValueChange,
-        modifier = modifier.testTag(CARD_NUMBER_FIELD),
+        modifier = modifier,
         labelText = stringResource(R.string.add_card_number_field_title),
         placeholderText = stringResource(R.string.add_card_number_field_hint),
         keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
@@ -45,7 +43,7 @@ fun CardNumberFieldPreview() {
         CardNumberField(
             value = text,
             onValueChange = {
-                text =  CardUiState().withCardNumber(it).cardNumber
+                text = CardUiState().withCardNumber(it).cardNumber
             },
         )
     }
