+++ app/src/main/java/woowacourse/payments/ui/features/addcard/components/CardPasswordField.kt
@@ -8,7 +8,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
@@ -17,7 +16,6 @@ import woowacourse.payments.R
 import woowacourse.payments.ui.components.AppTextField
 import woowacourse.payments.ui.features.addcard.CardUiState
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
-import woowacourse.payments.ui.util.AppTestTags.CARD_PASSWORD_FIELD
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -29,7 +27,7 @@ fun CardPasswordField(
     AppTextField(
         value = value,
         onValueChange = onValueChange,
-        modifier = modifier.testTag(CARD_PASSWORD_FIELD),
+        modifier = modifier,
         labelText = stringResource(R.string.add_card_password_field_title),
         placeholderText = stringResource(R.string.add_card_password_field_hint),
         keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password),
