--- app/src/main/java/woowacourse/payments/ui/util/AppTestTags.kt
@@ -1,9 +1,0 @@
-package woowacourse.payments.ui.util
-
-object AppTestTags {
-    const val CARD_NUMBER_FIELD = "card_number_field"
-    const val CARD_EXPIRY_DATE_FIELD = "card_expiry_date_field"
-    const val EXPIRE_DATE_ERROR_TEXT = "expire_date_error_text"
-    const val CARD_OWNER_NAME_FIELD = "card_owner_name_field"
-    const val CARD_PASSWORD_FIELD = "card_password_field"
-}
